pyproject.toml
[project]
name = "github-mlops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-storage-blob>=12.28.0",
    "fastapi>=0.128.1",
    "joblib>=1.5.3",
    "mlflow>=3.9.0",
    "numpy>=2.4.2",
    "pandas>=2.3.3",
    "plotly>=6.5.2",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "streamlit>=1.54.0",
    "uvicorn>=0.40.0",
]
